+++ src/main/java/com/aly/ecomapp/service/ProductService.java
@@ -214,7 +214,7 @@ public class ProductService {
         return dto;
     }
 
-    public List<ProductDto> getAllFilteredProducts(String title, Integer catId, Integer priceMin, Integer priceMax) {
+    public List<ProductDto> getAllFilteredProducts(String title, Long catId, Double priceMin, Double priceMax) {
     return productRepository.findAllByCondition(title, catId, priceMin, priceMax).stream()
             .map(this::convertToDto)
             .collect(Collectors.toList());
